requirements.txt
fastapi
uvicorn
twilio
python-multipart
pillow
pillow-heif
httpx
